refactoring api client: add network client, auth api client, image downloader, api client exceptions and configuration
