+++ lib/configuration/configuration.dart
@@ -1,0 +1,6 @@
+class Configuration {
+  // TODO: надо бы убрать static
+  static const host = 'https://api.themoviedb.org/3';
+  static const imageUrl = 'https://image.tmdb.org/t/p/w500';
+  static const apiKey = '16257858b79d98176712e6d572638798';
+}
