+++ lib/domain/api_client/api_client_exception.dart
@@ -1,0 +1,7 @@
+enum ApiClientExceptionType { network, auth, other, sessionExpired }
+
+class ApiClientException implements Exception {
+  final ApiClientExceptionType type;
+
+  ApiClientException(this.type);
+}
