+++ lib/domain/api_client/auth_api_client.dart
@@ -1,0 +1,73 @@
+import 'package:comics_db_app/configuration/configuration.dart';
+import 'package:comics_db_app/domain/api_client/network_client.dart';
+
+class AuthApiClient {
+  final _networkClient = NetworkClient();
+
+  Future<String> auth({
+    required String username,
+    required String password,
+  }) async {
+    final token = await _makeToken();
+    final validToken = await _validateUser(
+        username: username, password: password, requestToken: token);
+    final sessionId = await _makeSession(requestToken: validToken);
+    return sessionId;
+  }
+
+  Future<String> _makeToken() async {
+    String parser(dynamic json) {
+      final jsonMap = json as Map<String, dynamic>;
+      final token = jsonMap['request_token'] as String;
+      return token;
+    }
+
+    final result = _networkClient.get(
+      '/authentication/token/new',
+      parser,
+      <String, dynamic>{'api_key': Configuration.apiKey},
+    );
+    return result;
+  }
+
+  Future<String> _validateUser(
+      {required String username,
+      required String password,
+      required String requestToken}) async {
+    String parser(dynamic json) {
+      final jsonMap = json as Map<String, dynamic>;
+      final token = jsonMap['request_token'] as String;
+      return token;
+    }
+
+    final parameters = <String, dynamic>{
+      'username': username,
+      'password': password,
+      'request_token': requestToken
+    };
+    final result = _networkClient.post(
+      '/authentication/token/validate_with_login',
+      parameters,
+      parser,
+      <String, dynamic>{'api_key': Configuration.apiKey},
+    );
+    return result;
+  }
+
+  Future<String> _makeSession({required String requestToken}) async {
+    String parser(dynamic json) {
+      final jsonMap = json as Map<String, dynamic>;
+      final sessionId = jsonMap['session_id'] as String;
+      return sessionId;
+    }
+
+    final parameters = <String, dynamic>{'request_token': requestToken};
+    final result = _networkClient.post(
+      '/authentication/session/new',
+      parameters,
+      parser,
+      <String, dynamic>{'api_key': Configuration.apiKey},
+    );
+    return result;
+  }
+}
