+++ lib/domain/api_client/image_downloader.dart
@@ -1,0 +1,5 @@
+import 'package:comics_db_app/configuration/configuration.dart';
+
+class ImageDownloader {
+  static String imageUrl(String path) => Configuration.imageUrl + path;
+}
