+++ lib/ui/widgets/tv_details/tv_details_model.dart
@@ -1,4 +1,5 @@
 import 'package:comics_db_app/domain/api_client/api_client.dart';
+import 'package:comics_db_app/domain/api_client/api_client_exception.dart';
 import 'package:comics_db_app/domain/data_providers/session_data_provider.dart';
 import 'package:comics_db_app/domain/entity/tv_details.dart';
 import 'package:flutter/cupertino.dart';
